# backend/docs/swagger.json
{
    "components": {
        "schemas": {
            "auth.authResponse": {
                "properties": {
                    "access_token": {
                        "type": "string"
                    },
                    "expires_in": {
                        "example": 900,
                        "type": "integer"
                    },
                    "token_type": {
                        "enum": [
                            "Bearer"
                        ],
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/auth.userResponse"
                    }
                },
                "required": [
                    "access_token",
                    "expires_in",
                    "token_type",
                    "user"
                ],
                "type": "object"
            },
            "auth.credentialsRequest": {
                "properties": {
                    "email": {
                        "description": "Validate the address after normalization so surrounding whitespace remains accepted.",
                        "example": "demo@example.com",
                        "type": "string"
                    },
                    "password": {
                        "example": "correct horse battery staple",
                        "maxLength": 128,
                        "minLength": 12,
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password"
                ],
                "type": "object"
            },
...
            "auth.userResponse": {
                "properties": {
                    "avatar_url": {
                        "type": "string"
                    },
                    "created_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "email": {
                        "format": "email",
                        "type": "string"
                    },
                    "id": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "required": [
                    "avatar_url",
                    "created_at",
                    "email",
                    "id",
                    "name"
                ],
                "type": "object"
            },
...
            "httpx.ErrorResponse": {
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/httpx.ErrorDetail"
                    }
                },
                "required": [
                    "error"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "bearerFormat": "JWT",
                "description": "Access JWT issued by registration, login, or refresh.",
                "scheme": "bearer",
                "type": "http"
            },
            "RefreshCookie": {
                "description": "HttpOnly refresh cookie scoped to /api/v1/auth. Sessions expire 30 days after login; rotation does not extend expiry.",
                "in": "cookie",
                "name": "refresh_token",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "Authentication, profiles, avatars, and website monitor configuration. Unknown JSON fields are rejected. Monitoring checks are not implemented.\nRequests with Origin must match ALLOWED_ORIGIN. Mutating requests without Origin require X-CSRF-Protection: 1. Access JWTs expire after 900 seconds; logout does not invalidate existing JWTs.",
        "title": "Uptime API",
        "version": "1.0.0"
    },
...
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Accepts only email and password. Sets an HttpOnly refresh_token cookie. JSON body is limited to 4096 bytes. Email is trimmed and lowercased; password length is measured in Unicode characters.",
                "operationId": "login",
                "parameters": [
                    {
                        "description": "Allowed frontend origin; required for browser requests",
                        "in": "header",
                        "name": "Origin",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Must be 1 for mutating requests without Origin",
                        "in": "header",
                        "name": "X-CSRF-Protection",
                        "schema": {
                            "enum": [
                                "1"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/auth.credentialsRequest",
                                "summary": "credentials",
                                "description": "Email and password"
                            }
                        }
                    },
                    "description": "Email and password",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.authResponse"
                                }
                            }
                        },
                        "description": "Account and access token",
                        "headers": {
                            "Set-Cookie": {
                                "description": "HttpOnly refresh_token; SameSite=Lax; Secure when configured",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Invalid JSON or credentials"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Invalid credentials"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Origin or CSRF rejected"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Authentication capacity exhausted",
                        "headers": {
                            "Retry-After": {
                                "description": "1 second",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal server error"
                    }
                },
                "summary": "Sign in",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Idempotently revokes the current refresh session and clears its cookie. The refresh_token cookie is optional. Existing access JWTs remain valid until expiry.",
                "operationId": "logout",
                "parameters": [
                    {
                        "description": "Allowed frontend origin; required for browser requests",
                        "in": "header",
                        "name": "Origin",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Must be 1 for mutating requests without Origin",
                        "in": "header",
                        "name": "X-CSRF-Protection",
                        "schema": {
                            "enum": [
                                "1"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional refresh_token=\u003ctoken\u003e cookie; browsers send it automatically",
                        "in": "header",
                        "name": "Cookie",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Signed out; no response body",
                        "headers": {
                            "Set-Cookie": {
                                "description": "Deletes refresh_token",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Origin or CSRF rejected"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal server error"
                    }
                },
                "summary": "Sign out",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/me": {
            "get": {
                "description": "Returns the account identified by the Bearer JWT, including a public avatar URL or an empty string.",
                "operationId": "me",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.userResponse"
                                }
                            }
                        },
                        "description": "Authenticated account"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Invalid access token or deleted user"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Origin rejected"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get the authenticated account",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Uses the HttpOnly refresh_token cookie and replaces it on success. Refresh calls must be sequential; replay revokes the session. Session expiry is fixed and invalid cookies are cleared.",
                "operationId": "refresh",
                "parameters": [
                    {
                        "description": "Allowed frontend origin; required for browser requests",
                        "in": "header",
                        "name": "Origin",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Must be 1 for mutating requests without Origin",
                        "in": "header",
                        "name": "X-CSRF-Protection",
                        "schema": {
                            "enum": [
                                "1"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.tokenResponse"
                                }
                            }
                        },
                        "description": "New access token; no user field",
                        "headers": {
                            "Set-Cookie": {
                                "description": "Replacement HttpOnly refresh_token",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Missing, expired, revoked, or replayed refresh token",
                        "headers": {
                            "Set-Cookie": {
                                "description": "Deletes the invalid refresh_token cookie",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Origin or CSRF rejected"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "RefreshCookie": []
                    }
                ],
                "summary": "Rotate the refresh token",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/register": {
            "post": {
                "description": "Accepts only email and password. Sets an HttpOnly refresh_token cookie. JSON body is limited to 4096 bytes. Email is trimmed and lowercased; password length is measured in Unicode characters.",
                "operationId": "register",
                "parameters": [
                    {
                        "description": "Allowed frontend origin; required for browser requests",
                        "in": "header",
                        "name": "Origin",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Must be 1 for mutating requests without Origin",
                        "in": "header",
                        "name": "X-CSRF-Protection",
                        "schema": {
                            "enum": [
                                "1"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/auth.credentialsRequest",
                                "summary": "credentials",
                                "description": "Email and password"
                            }
                        }
                    },
                    "description": "Email and password",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.authResponse"
                                }
                            }
                        },
                        "description": "Account and access token",
                        "headers": {
                            "Set-Cookie": {
                                "description": "HttpOnly refresh_token; SameSite=Lax; Secure when configured",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Invalid JSON or credentials"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Origin or CSRF rejected"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Email already registered"
                    },
                    "429": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Authentication capacity exhausted",
                        "headers": {
                            "Retry-After": {
                                "description": "1 second",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal server error"
                    }
                },
                "summary": "Register an account",
                "tags": [
                    "auth"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "description": "Current API server",
            "url": "/api/v1"
        }
    ]
}